AI RESUME/requirements.txt
fastapi
uvicorn[standard]
python-multipart
scikit-learn
numpy
scipy
python-docx
PyPDF2
Jinja2
